refactor: Move getWaveDefinition to waveManager and clean up imports across modules
